rename `transforms` to `graph_transforms` with rotations and reflections
