+++ tests/models/test_cgcnn.py
@@ -14,7 +14,7 @@ import torch
 from ase.io import read
 from torch_geometric.data import Data
 
-from ocpmodels.common.transforms import RandomRotate
+from ocpmodels.common.graph_transforms import RandomRotate
 from ocpmodels.datasets import data_list_collater
 from ocpmodels.models import CGCNN
 from ocpmodels.preprocessing import AtomsToGraphs
