+++ tests/models/test_dimenetpp.py
@@ -14,7 +14,7 @@ import pytest
 import torch
 from ase.io import read
 
-from ocpmodels.common.transforms import RandomRotate
+from ocpmodels.common.graph_transforms import RandomRotate
 from ocpmodels.datasets import data_list_collater
 from ocpmodels.models import DimeNetPlusPlus
 from ocpmodels.preprocessing import AtomsToGraphs
